fix(test): jest.mockホイスティング説明コメント追加・en-mockヘルパーを共通化 #851
